+++ idea/testData/intentions/convertSealedClassToEnum/dropDefaultConstructorCall.kt
@@ -1,0 +1,3 @@
+sealed class <caret>MyClass {
+    object FOO : MyClass()
+}