+++ idea/testData/intentions/convertSealedClassToEnum/outOfRange.kt
@@ -1,0 +1,5 @@
+// IS_APPLICABLE: false
+
+<caret>private sealed class X {
+
+}